+++ R/sigminer.R
@@ -19,3 +19,13 @@
 #' @import cli
 #' @name sigminer
 NULL
+
+## usethis namespace: start
+#' @useDynLib sigminer, .registration = TRUE
+## usethis namespace: end
+NULL
+
+## usethis namespace: start
+#' @importFrom Rcpp sourceCpp
+## usethis namespace: end
+NULL